Refactor native features into capability modules
