+++ crates/tryke_discovery/Cargo.toml
@@ -10,8 +10,8 @@ authors.workspace = true
 license.workspace = true
 
 [features]
-default = ["native"]
-native = [
+default = ["filesystem"]
+filesystem = [
   "dep:ignore",
   "dep:log",
   "dep:rayon",
