+++ crates/tryke_reporter/Cargo.toml
@@ -10,8 +10,8 @@ authors.workspace = true
 license.workspace = true
 
 [features]
-default = ["native"]
-native = [
+default = ["terminal"]
+terminal = [
   "dep:clearscreen",
   "dep:console",
   "dep:ctrlc",
