+++ crates/tryke_reporter/src/clear/fallback.rs
@@ -1,0 +1,6 @@
+#[must_use]
+pub fn stdout_is_terminal() -> bool {
+    false
+}
+
+pub fn clear_terminal() {}
