+++ crates/tryke_reporter/src/lib.rs
@@ -7,7 +7,7 @@ pub mod junit;
 pub mod live;
 pub mod llm;
 pub mod next;
-#[cfg(feature = "native")]
+#[cfg(feature = "terminal")]
 pub mod progress;
 pub mod reporter;
 pub mod sugar;
@@ -19,7 +19,7 @@ pub use json::JSONReporter;
 pub use junit::JUnitReporter;
 pub use llm::LlmReporter;
 pub use next::NextReporter;
-#[cfg(feature = "native")]
+#[cfg(feature = "terminal")]
 pub use progress::ProgressReporter;
 pub use reporter::Reporter;
 pub use sugar::SugarReporter;
