+++ crates/tryke_reporter/src/live/mod.rs
@@ -1,0 +1,9 @@
+#[cfg(not(feature = "terminal"))]
+mod fallback;
+#[cfg(feature = "terminal")]
+mod terminal;
+
+#[cfg(not(feature = "terminal"))]
+pub use fallback::*;
+#[cfg(feature = "terminal")]
+pub use terminal::*;
